requirements.txt
reflex==0.6.6
sqlalchemy==2.0.30
# psycopg2-binary==2.9.9
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
openpyxl==3.1.2
Pillow==10.3.0
python-dotenv==1.0.1
